requirements.txt
appdirs
Django
djangorestframework
django-rest-framework-csv
markdown
django-filter
packaging
psycopg2
pyparsing
six
unicodecsv
